requirements.txt
openai==1.12.0
httpx==0.24.1
flask==3.0.3
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
requests==2.32.3
